ido-style return on directories
I would like return to perform the persistent action on directories instead of opening the directory.  

This allows this.  I would like to resume the helm session and then perform the persistent action on the candidate instead, but I'm not familiar enough with the inner workings of helm to set this up.