More elaborated tracker-search source
Working towards multiline candidates 'helm-source-tracker-transformer can
handle (DISPLAY . REAL) cells and builds them now.

To make that possible, a new attribute 'source-candidate-seperator (not
yet documented) is queried in 'helm-output-filter--process-source. It
defines the 'split-string SEPERATOR by which the raw output source is
split. This is needed to make the raw candidates arrive in one piece in
form of
"  file:///path/to/foo
  ...foo snippet..."