A function to enumerate a directory recursively.
This function is taken from https://gist.github.com/dmgerman/5675462. I
added only a helm- prefix to follow the convention for helm
functions. This function will be used in a subsequent commit to help
provide a helm source that provides files from a configured set of
directories, with an option to process each directory recursively.