Bug fix: Non-string value for REAL with multi-select action.
An error was raised when one use non-string value for REAL of
(DISPLAY . REAL) pair and try to execute command with multiple
candidates.  The cause is `string=` was used for equivalence
checking of REAL.

Note that using `equal` is not the best way.  If we use an object
containing circular list for a value of REAL, we fall into infinite
loop.