Escape regex characters in `helm--prompt`
I ran into this issue after the "->" was added because helm-projectile uses "[project] pattern:" for it's prompt.  When `string-match` does this comparison, the "[project]" part is treated as a set of characters rather than a raw string, thus always failing the condition in the `if`.

I've changed the test to treat `helm--prompt` as a raw string as well as changed it to use `string-match-p` since you aren't using the returned index.