Control insertion of kill-ring entry into buffer.
Add a defcustom to control whether an entry selected from
`helm-show-kill-ring` gets inserted into the buffer.  By default,
nothing is different for the user, but when the new defcutsom
`helm-show-kill-ring-yank` is set to nil, selecting an entry from
`helm-show-kill-ring` will not cause that entry to be inserted into the
buffer.  This is useful for users who are using, e.g., clipmon-mode to
synchronize the kill-ring and the X clipboard.  If the user wants to
choose an entry from the kill ring to paste into another application,
s/he can choose it with `helm-show-kill-ring`, then switch to the other
application to paste.