Fix #1451 by storing external commands with paths.
Now, it no longer requires two attempts to open a file of a particular
type externally.

`helm-external-commands-list` now stores commands with fully qualified
paths (e.g. `/usr/bin/xdg-open` instead of `xdg-open`). The only
exception to this is user-invoked commands (through interactive
`helm-run-external-command`). For a good user experience, we should not
qualify the commands the user sees with paths.

Various functions (e.g. `helm-run-and-raise`) no longer require the
command to be a format string (e.g. `/usr/bin/xdg-open %s`). They now
take the command by itself (e.g. `/usr/bin/xdg-open`). This eliminates
several repetitive calls to `replace-regexp-in-string`. This should not
cause any problems, because we were already assuming that the commands
are in the form `COMMAND %s`.