delay highlighting file-attribute retrieval
This is a change that was discussed heavily in #749, fixed
incorrectly in c26d3a9 (this change ignored whether the file
in question was remote or not) and then essentially reverted
in 46687d1.

Here is the correct fix and speeds up my heavy TRAMP usage a ton
(from ~5 seconds pulling up helm-mini to instantaneous).

In order to recreate this, populate your recentf with remote files,
open one of these remote files (this part is 100% necessary), and
then do a helm-mini.  Since the connection to the remote files is open
the file-attribute will retrieve all the attributes of the files.

* helm-files.el (helm-highlight-files): delay file-attribute check