Fix helm-man-woman in failures of man and woman
Changes:
- Pass absolute man file path to `man' without -l flag.
- When woman fails, the selected man section is retained. This fixes
the undesired behavior that only the topic is passed to `man' after
such a failure.

The rationale behind the first change is that in BSD, man has no -l
option. Since the file name passed to man is generated from a woman
helper function, it is extremely unlikely to resemble a page
name. (see Help on `man')