Remove timer in `helm-info-lookup-symbol`.
For now it seems everything is working fine without timer, also with
the timer we loose all keymaps when switching to the source with
several choices for info when started from `helm-cycle-resume'
but not from `helm-run-cycle-resume' and normal resume.  If
everything work fine without timer `helm-info-lookup-symbol-1' is
no more needed.

* helm-elisp.el (helm-info-lookup-symbol): Do it.