helm.el (helm-update): Display "results not found" if no matches (#1806)
This is a proof-of-concept that displays "Results not found." if there
are no matches found for any sources in the Helm session and the input
is empty.  It may not be the best way, and it hasn't been thoroughly
tested (e.g. with async sources), but it seems to work for sync
sources.  Maybe it is a starting point.