helm-org.el (helm-org-refile-within-buffer): Add function
This function refiles the entry at point to a heading selected with Helm
within BUFFER.  Interactively, BUFFER defaults to the current buffer.

When you want to refile within the current buffer, this is much faster
than org-refile, because it only looks at headings in the current
buffer.  If you have a lot of headings in a lot of open files,
org-refile can take several seconds to bring up the list of headings in
all files.

This was partially inspired by John Kitchin's post at
<http://kitchingroup.cheme.cmu.edu/blog/2016/03/18/Org-teleport-headlines/>