helm-org.el (helm-org-in-buffer-headings): Don't nullify helm-org-format-outline-path
Since this command shows headings in the current buffer, it makes sense
to hide the filename, but showing the full outline path can be very
useful in narrowing down the list.  For example, in a tree like:

* Languages
** English
*** Alphabet
** Greek
*** Alphabet

If the user wanted to go to the "Greek/Alphabet" heading, without
displaying the outline path, he'll see this after typing "alphabet":

*** Alphabet
*** Alphabet

Which wouldn't be useful.  But with full paths, he'd see:

** English/Alphabet
** Greek/Alphabet

By leaving helm-org-format-outline-path alone, we respect the user's
setting, and allow it to be overridden by binding it around this command.