Fix ucs-name handling for upcoming Emacs release
A new incompatible change is that ucs-names will be a hash-table
instead of a list.  See Emacs Bug#28302.

* helm-font.el (helm-calculate-ucs-alist-max-len): Add function to
calculate ucs-names candicates when wrapped inside a list.
(helm-calculate-ucs-hash-table-max-len): Add function to calculate
ucs-names candicates when wrapped inside a hash-table.
(helm-calculate-ucs-max-len): Refactor to support the hash-table.
(helm-ucs-collect-symbols-alist): Add function to collect the
ucs-names list.
(helm-ucs-collect-symbols-hash-table): Add function to collect the
ucs-names hash-table.
(helm-ucs-collect-symbols): Refactor to support the hash-table.