Pull out the logic for  pipe commands into a separate function.
Having a separate function allows the user to advise and/or re-define
just that function if they let-bind `helm-grep-default-command' around
a call to `helm-do-grep-1' and friends, but still want to have the
"space-separated patterns pipe to grep" functionality.  (In my case, I
let-bound it to command that runs codesearch to make a quick-and-dirty
Emacs interface for that command).