Allow toggling full frame (#1931)
Introduce a new slot for sources, :persistent-action-if.
It allows with one function defining several functions that will be
used and will behave differently for splitting helm-window under
certain conditions.

* helm.el (helm-map): Bind it.
(helm-toggle-full-frame): New.
(helm-execute-persistent-action): Now the optional ATTR arg can be
persistent-action or persistent-action-if.
* helm-files.el
(helm-find-files-persistent-action-if): New, replace
helm-find-files-persistent-action.
(helm-source-ffiles): Use PA-if instead of PA.