Add a slash instead of a space when point is behind a directory
Pressing <tab> when point is behind a directory that does not contain
a / will result in a space being added. If the user wanted to continue
tabbing they would need to backspace, type a / and then they can
continue pressing / for completions.

This is different behaviour to what one might encounter in other
shells or in eshell with the default completion. This change
reintroduces that behaviour by adding a / instead of a space.