Avoid excessively calling format-mode-line
During the initial buffer list setup, `format-mode-line` and `length`
were called in every buffer regardless of the values of
`helm-buffer-max-len-mode` and `helm-buffer-max-length`. Since the
performance of format-mode-line depends on the mode, this can really
slow down the initialization of the buffer list. For example,
`delight-mode`'s advice around `format-mode-line` was causing
an obscene number of calls to `dired-get-filename`.

format-mode-line         1044        9.8059119999  0.0093926360
helm-buffers-list--init  1           9.804546      9.804546
dired-get-filename       120960      0.0           0.0

Might as well only do this computation once.