Fix a bug with prefixed fnames in hff
where pattern is matching a file but is at the same time equal to
another existing file, e.g. ack and ack-grep in /usr/bin.

* helm-files.el (helm-find-files-get-candidates): Check earlier (here)
if pattern is an existing file.
(helm-ff-sort-candidates-1): move test of "pattern is an existing file"
to helm-find-files-get-candidates as is is too late to do it here for
remote files.