Fix window selection with multiple side windows
While trying to debug the follwing issues:
- https://github.com/emacs-helm/helm/issues/2120
- https://github.com/syl20bnr/spacemacs/issues/11368

Having:
```
(helm-default-display-buffer-functions '(display-buffer-in-side-window))
```
I ended up in the `helm-persistent-action-display-window` function.

When no other side windows exist here is the list of windows:
```
(window-list): (#<window 8 on *helm find files*> #<window 2 on  *Minibuf-1*> #<window 1 on *Messages*>)
(selected-window): #<window 8 on *helm find files*>
(next-window (selected-window) 1): #<window 1 on *Messages*>
(previous-window (selected-window) 1): #<window 1 on *Messages*>
```

When a side window already exists, for example treemacs:
```
(window-list): (#<window 12 on *helm find files*> #<window 2 on  *Minibuf-1*> #<window 10 on  *Treemacs-Framebuffer-1*> #<window 1 on *Messages*>)
(selected-window): #<window 12 on *helm find files*>
(next-window (selected-window) 1): #<window 10 on  *Treemacs-Framebuffer-1*>
(previous-window (selected-window) 1): #<window 1 on *Messages*>
```

Notice the treemacs window is right after the selected window, not
including the minibuffer, and the window we actually want to check if
it's dedicated or not (window 1) is right at the end.

Using the `previous-window` instead of `next-window` ensures that we
indeed get to check the window we care about (window 1). In the first
place it happend to be the case that both functions returned the same
(correct) value. Also `previous-window` makes sense since `window-list`
behaves like a ring and switching from a normal window to a helm window
means the normal window gets "left behind" in the previous position
while the "next window" would be whatever window was ever touched last.