Allow editing rsync command
May be necessary when we want to add extra arg like --delete.

When a prefix arg is passed, setup the rsync switches in
helm-find-files-do-action and pass this new value
to helm-rsync-copy-files with its new optional arg SWITCHES.

When the prefix arg is passed user is prompted for modifying rsync command.