Handle Straight user config
Make was not working with async installed in Straight, and emacs-helm.sh as
well.

In Makefile try to find Straight build directory and add it to LOAD-PATH if
found.
Same for emacs-helm.sh and don't use package.el if straight directory is
found.

Now Helm compile with async installed with Straight and emacs-helm.sh is
working with no extra config.