Fix running a command involving helm from M-x
When running a command involving again helm from helm-M-x, the
unwind-protect UNWINDS forms are executed only once this helm
command exit leaving the helm-M-x timer running and other variables
and hooks not unset, as a result the prefix arg continues to be shown in next
helm command.

So the timer is now in a global var and all
the forms that should normally run in unwind-protect are running as soon as
helm-M-x-execute-command is called.