Cache candidates in HFF
Each time HFF enter a new directory its contents is cached, further visits to
this directory use the cache, deleting a file in PA or force updating delete
the helm-ff-default-directory entry in cache.

The cache creation and update is done in helm-list-directory, the source have
a update function that clear the cache when force updating, finally the cache
is cleared at end of eache HFF session.