Cache the filtered candidates
Add the text properties directly to the result of helm-ff-directory-files and
cache all instead of running a FCT on top of cached candidates.
Now the filters are returning a list of strings instead of a list of cons
cell (display . real), each string have a helm-realvalue prop.
Now use only one simple FCT that display basename or fullpath.
Tramp hostnames are filtered separately, same for the invalid fnames on
top (still not working though).

FIXME:
- Invalid fnames are not treated.
- local directory variable are not handled.