Use match-on-real in HFF
From now get rid of helm-ff-transformer-show-only-basename.
All candidates are now stored as cons cell (display . real) in hash-table,
where display is the basename and real the full path of file.
With this there is no need of constantly looping in candidates to match on
real name and then display helm buffer with basenames, but because the pattern
is always a full path we can't match anymore candidates which are displayed
directly as basenames hence the use of match-on-real.