Allow using eshell cmd on file async (fix emacs bug)
This fix avoid an error with format, and allows returning the correct value
expected by eshell-command to setup and eshell async buffer.

This is done by advicing temporarily `eshell-eval-command` which should return
a list of one element _a process_ when the command ends with "&".

Now users can safely run commands ending with "&", adding "&" to aliases still
impossible though for now.