Handle nil value for helm-buffers-sort-fn
...or for helm-fuzzy-matching-highlight-fn.  By default,
helm-buffers-sort-fn is initialized with the value of helm-fuzzy-sort-fn,
and the documentation string for helm-fuzzy-sort-fn states that it may be
set to nil to indicate that no sorting should be done.  Similarly, the
documentation string for helm-fuzzy-matching-highlight-fn states that it
may be set to nil to indicate that no highlighting should be done.  Thus
functions that use helm-buffers-sort-fn or helm-fuzzy-matching-highlight-fn
to sort or highlight a list of candidates must handle nil values for these
options by using the candidates directly.

* helm-buffers.el (helm-highlight-buffers): Handle nil value for
helm-fuzzy-matching-highlight-fn.
(helm-buffers-sort-transformer): Handle nil value for helm-buffers-sort-fn.
* helm.el (helm-fuzzy-highlight-matches): Handle nil value for
helm-fuzzy-matching-highlight-fn.