New variable to allows completing-read return a list
When non nil `helm-comp-read' will return marked candidates, a list instead of
a single candidate, even if no mark it will return a list of one elemnt, the
selected candidate.
This will break all completing-reads when helm-mode is enabled if set globally,
have to be only used in let bindings, this allows external packages to make
use of marks in their completing-reads when helm-mode is enabled.