Add filters to show files or dirs only in HFF
This is done by adding two filtered-candidate-transformer functions and two
toggle functions that change the value of the corresponding flags.
A third function have been added to restore to "show all".
Show all after moving out of the directory and also when quitting or exiting
(flags are reset to nil).