Try again to fix error in emacs-28 with selected-frame
AFAIU, selected-frame should be called from helm-window to get the right frame
i.e. what we are looking for is the frame displaying helm-buffer but when for
example in shell completion the window (in frame) is never created when
exiting with helm-execute-action-at-once-if-one so fallback to selected-window
to avoid error with with-selected-window called with a nil parameter.