Fix helm-this-command for emacs-28
This change tries to fix a serie of problems introduced in emacs-28 which
doesn't happen in emacs-27, BTW it may be reverted if it cause problems in
emacs-27.

For some reasons we have to exclude helm-M-x, otherwise helm-this-command goes
up to helm-M-x instead of returning the command executed by helm-M-x.
This doesn't happen on emacs-27.

To fix it, helm-M-x is added to the ignored function list (now a defconst) and
backtrace-frame is starting from this-command.