Bind inhibit-quit in helm-read-from-minibuffer
As we are using `helm-keyboard-quit' for `C-g' we have to
prevent emacs command loop redefining `C-g' during
helm-session. This happen only on async source with large
output after a certain delay. The effect is that the minibuffer
is exited but the helm async process continue running, because
minibuffer is lost `C-g' have no more effect. By binding
`inhibit-quit' here we prevent this and allow `C-g' (the helm
one aka `helm-keyboard-quit') to quit immediately.