Set all windows dedicated to non-dedicated on start (#2243)
Until now we had code in helm-persistent-action-display-window to deal
with dedicated windows, it was fixing bug #2050 and was working
reasonabily well, however now we have popular applications like lsp,
treemacs and whatever abusing of dedicated windows we need a more
drastic fix.

All dedicated windows are now set to non-dedicated on startup and
restored when exiting.