Fix arguments error on read-coding-system
read-coding-system calls completing-read-default, but is not
interactive, and not recognized by helm-this-command.

If e.g. write-region would call read-coding-system due to incompatible
bytes in the buffer, helm-this-command would recognize write-region
further up the stack, and pick the associated
helm-read-file-name-handler-1 from helm-completing-read-handlers-alist,
which fails due to wrong argument count.

Make helm-this-command recognize read-coding-system, so the default helm
handler is used.