Prevent image-dired corrupting image-dired directory in emacs-29
image-dired is now using sha1 to build thumbnails filenames, as a
result old thumbnails files made with md5 are ignored and recreated
with sha1.

Force usage of md5 to fix this by redefining
`image-dired-get-thumbnail-image` and `image-dired-thumb-name`.
As a bonus cache thumbnails to not recompute the md5 each time we
revisit the directory or toggle thumbnails view.