Provide icons in helm-filtered-bookmarks
In addition to provide icons, the tricky part is to make adaptive
sorting work with the leading space where the icon is stored as a
diplay prop.
Also helm adaptive was storing the full display part in
helm-adaptive-history which was wrong when bookmark was shown with its
location; To fix it, now bookmark show its location in another display
prop stored in a trailing space.
helm adaptive is now removing both leading and trailing space when
comparing candidates.