Update helm-this-command for emacs-28
Before emacs-28, helm-this-command detects the command by the stack
frame information of call-interactively, even while the command does
not exist in backtrace-frames.
Typically, this happens in an interactive form.

In emacs-28, backtrace-frames in an interactive form does not have the
frame of call-interactively for the command, and a backtrace looks
like:
...
(descriptor of interactive form)
command-execute(command)
...
Consequently, helm-this-command falls into backtrace-frames and may
reach helm-M-x, execute-extended-command, or other M-x commands.

Even in emacs-28, the frame of command-execute still keeps the command
as an argument.
Added forms into helm-this-command check this information, and if it
is detected, return it.