Implement diacritics for in buffer sources
Diacritics was only implemented on sync source via the :match slot, as
it is problematic to implement it through this slot in in buffer
sources, we use a specific :diacritics slot in helm-source-in-buffer.

1) Implement matching on match-part with diacritics (helm-core).
2) Make helm-mm-3-search-base aware of diacritics and create a
specific fn for this.
3) Add diacritics slot in helm-source-in-buffer and prepare search
function when the feature is enabled though this slot.
4) Enable it on helm-occur for testing (will make it configurable and
nil by default later).
5) Fix order of match functions when diacritics is enabled.