Ensure tramp read persistent data in emacs-helm.sh
Without this tramp prompt constantly for password each time an input
is detected.
With this change, even with no entries in auth-source file tramp
prompt only once for password at start.
Also set auth-source to NOT ask to save password for this connection
in auth-source file (and don't do it), we don't want to have this file
edited at all.