Try to fix issue #2504 in eshell completion
This fix the default eshell completion based on completion-at-point
which fails to complete a filename not preceded by a meaningful
command like cd or ls.
Pcomplete and helm-eshell are working properly in this case.

Problem is that the base-size returned by completion-all-completions
in this case is always zero (seems it assume we are trying to complete
a command even if the metadata specify a file completion)

To fix it we try here to recalculate the base-size from the last / found.