Disable helm-define-key-with-subkeys when in kbd macro
Entering the loop is fine while defining kbd macro as long as the next
command do not use a minibuffer, if it is the case this next command
will not run properly because the minibuffer is in use.

To fix this only run COMMAND and exit in such cases.

Update docstring as well.