Setup helm-pattern in helm-initial-setup
Previously we were setting helm-pattern according to default in
helm-initial-setup and we were overriding it later in helm-initialize
(after calling helm-initial-setup).
Now we are setting it in the same place which make the code more
readable and comprehensive.
Also pattern is set before calling init functions in case inout or
default is used in these functions to initialize candidates.
See (#2530) and <https://github.com/emacs-helm/helm-mu/issues/54>.