Add macros for creating interactive commands from actions
helm-make-command-from-action for commands exiting helm.
helm-make-persistent-command-from-action for persistent commands.

Use it helm-files commands.

Fix helm-ff-toggle-basename
No need to call it with a CANDIDATE arg as it is not an action.
Fix preselection with icons.

Fix toggle update in HFF
It is not an action, no need to use candidate arg.