Stop removing inotify watchers from the callback
Removing the watcher from its callback call another time its callback
unexpectedly with a stopped event and delay the next event when
modifying files in a loop. As a result, when entering the last
directory of the loop, this one is not updated as it should, it is
updated however but much longer after.

So from now, remove the watcher only if it is invalid and only when
visiting its directory. This is much faster as well, avoiding unneeded
operations, i.e. no need to remove and add again a watcher when it is
valid, just leave it.