Allow overriding REQUIRE-MATCH completing-read arg for a specific
function.

In Emacs-29+ describe-function completing-read uses 'confirm as value
for its require-match arg for allowing user to describe a function not
already defined i.e. an advice defined before the function is defined.
As a result we end up with an unknow candidate on top and user needs
to go to next line to avoid pressing RET on such annoying candidate.
Anyway describe-function is matching this unknown function even with
require-match = t and allow pressing RET on it to have the
description.

We use now a new var helm-comp-read-require-match-overrides to allow
overriding require-match arg if needed.

See Emacs bug#64902.