Use the right affixation function in completion
We were previously using helm--symbol-completion-table-affixation
which was the same as help--symbol-completion-table-affixation but for
helm. This was ok at the time when only symbol completion were using
this, i.e. it was the only affixation fn in Emacs, now affixation are
used everywhere so ensure we call the affixation function decided by
the caller.