Allow affixation functions to return a function instead of a list
Normally affixation functions use COMPLETIONS as arg, and return a list of
modified COMPLETIONS. Now we allow in Helm affixations functions to return a
function instead, just like annotation functions. The function should return a
list of three elements like (comp prefix suffix). This increase significantly
the speed avoiding one useless loop on complete list of candidates.

Speed is visible specially when using emacs as emacs style, with helm
styles it is already very fast so the difference is not really visible.