Be consistent with candidate-transformer in both in-buffer and async
When candidate-transformer is specified in source ALL candidates should
be computed with the candidate-transformer function (in contrast with
filtered-candidate-transformer).  This to be consistent with what sync
sources do. The car of the cons is used for initial fetching of
candidates whereas the cdr is used after when searching (in this case
the candidate number limit is used).