Fix error with read-multiple-choice
Some commands like `kill-buffer' involve another function
involving a completing-read, in this case we want to stop at this
function and not go up to the initial interactive call (in this
case kill-buffer).

As the same problem may occur in other places, add a new defconst to
store such functions.

The issue is fixed by making helm-this-command returning the function
(non interactive) calling completing-read instead of the initial command.