Stop trying to find default arg from prompt
As it is a common practice in Emacs, it may happen often and may be
wrong in some cases where what specified in prompt differ from the
real default provided later in function and not as DEFAULT arg
(which is unreachable for us).
So keep failing to find DEFAULT and provide a Emacs bug when this
is reported.