Ensure OVERLAY is a ref-cell
This try to fix (blindly) an error happening in Emacs-31 (setting
constant nil...).  I suspect the error comes from here but I have no
clue yet and I wonder why the argument OVERLAY should be nil.
The error is rare and difficult to reproduce, when it happen once, it
doesn't happen anymore afterward.