Fix switching to a buffer not selected in a tab
Previously we could do this only when switching to a buffer in other
window with Helm.  If user have for example two windows and enable
tab-bar-mode manually the two windows were not took in account in the
buffers of the initial tab (emacs bug) unless
tab-bar-tab-name-function set to tab-bar-tab-name-all.

So we are now setting tab-bar-tab-name-function in a :set function in
helm-buffers-maybe-switch-to-tab.

As a result we can now switch to a buffer not current in its tab.

This fix partially issue #2684.