helm--maybe-process-filter-one-by-one-candidate is now a function
With native-compiler and Emacs-30+ we may have an error (attempt to set
constant...). This never happened in previous Emacs since this was
written (around emacs-24) but it happened a few time recently with
Emacs-31.0.50 (very rare).

So let's use a function instead which is cleaner anyway.

Avoid transforming CANDIDATE by side effect from the macro, instead
return the new value of function(s) called on it and set manually
CANDIDATE to this value in the caller.