Helm-Icons: Add central file to manage icon support
Helm-icons provides provider independent icon support for Helm.
Each provider can be added to the abstraction logic with small
changes.
The only requirement is that the API is mostly the same as
all-the-icons or nerd-icons.

Where possible use the provider specific handler functions such
icon-for-mode or icon-for-file are used.

For actions there are no specific functions so there's a separate
wrapper for that between the supported providers. The wrapper
can also where possible use a better matching icon where available
in a provider.

The idea is to centralize the icon enabler and then individually
have the option to disable the icon support for each source.
Currently this doesn't work exactly.