Better fix for retrieving last expanded symlink in HFF (commit 7f602741)
If we concat icon+fname+->+truename, we can't retrieve anymore last
expanded.
If we add a display prop on whole candidate, we can retrieve last
expanded but then helm-fuzzy-highlight-matches is confused.
So we now concat icon+fname+a space containing "->+truename" and we
search fname maybe ending with a space at eol.