Better fix for issue #2709 by using coerce
recentf-list was corrupted by recentf itself when passing filename
with properties to find-file (action) and the filename passed to
action (real candidate) was corrupted with text props when it was
equal to the display i.e. (abbreviate-file-name "/foo")=> "/foo".
So we cleanup the candidate(s) before passing it/them to action.