Add a specialized match fn for files
Use it in HFF and read-file-name.

When matching a filename like "/home/you/github/foo-bar.txt" there is no
problems with `helm-mm-3-match' as long as one of the patterns doesn't match the
basedir of filename but as soon as you try to match a file with a name matching
basedir we would match all the file of directory instead of just the files we want
e.g. with a pattern like "/home/you/github/ foo git" you would match all files
of directory instead of matching only "/home/you/github/foo-git.el"
because "git" will always match "github".