Avoid deleting and creating timer needlessly
Previously we cancelled the timer and initialized it again at each
call of move-selection, now we create a timer that cancel itself and
check at each call if the timer is alive before creating a new one.

As a result, for example when keeping C-n pressed and passing e.g. 10
candidates, previously 10 timers were cancelled and created, now only
one timer is created.