Fix #2741 - error when quitting helm grep ag
It seems that on MacOS delete-process behave differently according to
the value of process-connection-type.  When it is non nil (pty) it seems
delete-process behave like if the call was interactive and delete the
process-buffer which is helm-buffer, then when the sentinel is called
the helm-window doesn't exists anymore and we have an error with
(select-window nil).

Binding process-connection-type to nil seems to fix the issue.